day 15/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
assemblyai
google-generativeai
Jinja2
python-multipart